+++ docs/changes.md
@@ -3,6 +3,7 @@
 2021/06/03
   * 完善 scroll bar。
   * 完善 menu layout。
+  * 修改ubjson\_writer\_write\_key\_len的问题（感谢攀峰提供补丁）。
 
 2021/06/02
   * 优化 widget\_set\_visible（感谢智明提供补丁）